Refactored code to store temperature as int but use temperature as float in the rest of the code

+++ Core/Inc/Measurement/AHT20/aht20_driver.h
@@ -10,8 +10,8 @@
 /* ----------------------------------------------------------------------------------- */
 
 typedef void (*fp_init_AHT20)(void);
-typedef int (*fp_get_temp)(void);
-typedef int (*fp_get_hum)(void);
+typedef float (*fp_get_temp)(void);
+typedef float (*fp_get_hum)(void);
 
 typedef struct AHT20_t
 {
